src/exporters/word_exporter.py
"""
Word document exporter for control profiles
"""

from typing import Dict
from docx import Document


class WordExporter:
    """Export profile data to Word document format"""

    def __init__(self, profile_data: Dict):
        self.profile_data = profile_data

    def export(self, output_path: str):
        """Export to Word document file"""
        # TODO: Implement Word export
        pass
